internal/registry/docker/docker_test.go
package docker

import (
	"context"
	"fmt"
	"net/http"
	"os"
	"os/exec"
	"testing"

	"github.com/stretchr/testify/require"

	"github.com/valist-io/registry/internal/core/mock"
	"github.com/valist-io/registry/internal/core/types"
)

func TestPublish(t *testing.T) {
	ctx := context.Background()

	tmp, err := os.MkdirTemp("", "test")
	require.NoError(t, err, "Failed to create temp dir")
	defer os.RemoveAll(tmp)

	client, err := mock.NewClient(tmp)
	require.NoError(t, err, "Failed to create mock client")

	orgName := "valist"
	orgMeta := &types.OrganizationMeta{
		Name:        "Valist, Inc.",
		Description: "Accelerating the transition to web3.",
	}

	repoName := "sdk"
	repoMeta := &types.RepositoryMeta{
		Name:        "docker",
		Description: "Valist core sdk.",
		ProjectType: "docker",
		Homepage:    "https://valist.io",
		Repository:  "https://github.com/valist-io/valist",
	}

	createEvent, err := client.CreateOrganization(ctx, orgMeta)
	require.NoError(t, err, "Failed to create organization")
	orgID := createEvent.OrgID

	_, err = client.LinkOrganizationName(ctx, orgID, orgName)
	require.NoError(t, err, "Failed to link organization name")

	_, err = client.CreateRepository(ctx, orgID, repoName, repoMeta)
	require.NoError(t, err, "Failed to create repository")

	registryAddr := "valist.local:5000"
	registryPath := "valist.local:5000/valist/docker"

	go func() {
		err := http.ListenAndServe(registryAddr, NewHandler(client))
		require.NoError(t, err, "Failed to start http server")
	}()

	err = exec.Command("docker", "build", "--tag", registryPath, "./testdata/").Run()
	require.NoError(t, err, "Failed to build docker image")

	err = exec.Command("docker", "push", registryPath).Run()
	require.NoError(t, err, "Failed to push docker image")
}
